Move xAsprintf, xSnprintf and xStrdup to StringUtils.h
