+++ CPUMeter.c
@@ -10,6 +10,7 @@ in the source distribution for its full text.
 #include "CRT.h"
 #include "Settings.h"
 #include "Platform.h"
+#include "StringUtils.h"
 
 #include <assert.h>
 #include <stdlib.h>
