+++ ColorsPanel.c
@@ -9,6 +9,7 @@ in the source distribution for its full text.
 
 #include "CRT.h"
 #include "CheckItem.h"
+#include "StringUtils.h"
 
 #include <assert.h>
 #include <stdlib.h>
