+++ LoadAverageMeter.c
@@ -9,6 +9,7 @@ in the source distribution for its full text.
 
 #include "CRT.h"
 #include "Platform.h"
+#include "StringUtils.h"
 
 
 static const int LoadAverageMeter_attributes[] = {
