+++ TasksMeter.c
@@ -7,8 +7,9 @@ in the source distribution for its full text.
 
 #include "TasksMeter.h"
 
-#include "Platform.h"
 #include "CRT.h"
+#include "Platform.h"
+#include "StringUtils.h"
 
 
 static const int TasksMeter_attributes[] = {
