+++ zfs/ZfsCompressedArcMeter.c
@@ -10,6 +10,7 @@ in the source distribution for its full text.
 
 #include "CRT.h"
 #include "Platform.h"
+#include "StringUtils.h"
 
 #include <stdlib.h>
 #include <string.h>
